requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-chroma
langchain-huggingface
sentence-transformers
langchain-groq
pypdf
python-dotenv
google-genai